requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
pytest==7.4.3
pytest-asyncio==0.21.1